+++ python/GafferUI/CompoundEditor.py
@@ -180,11 +180,7 @@ class CompoundEditor( GafferUI.Editor ) :
 
 		scriptWindow = self.ancestor( GafferUI.ScriptWindow )
 		if scriptWindow :
-			panel.setTitle( scriptWindow.getTitle() )
-			weakSetTitle = Gaffer.WeakMethod( panel.setTitle )
-			panel.__titleChangedConnection = scriptWindow.titleChangedSignal().connect( lambda w, t : weakSetTitle( t ), scoped = True )
-			# It's not directly in the qt hierarchy so shortcut events don't make it to the MenuBar
-			scriptWindow.menuBar().addShortcutTarget( panel )
+			panel.connectToScriptWindow( scriptWindow )
 
 		self.__detachedPanels.append( panel )
 		return panel
@@ -199,6 +195,16 @@ class CompoundEditor( GafferUI.Editor ) :
 		assert( not panel.visible() )
 		GafferUI.WidgetAlgo.keepUntilIdle( panel )
 
+	def _displayTransformChanged( self ) :
+
+		GafferUI.Editor._displayTransformChanged( self )
+
+		# Propagate display transform to detached panels. This would
+		# happen naturally if they were parented to the main ScriptWindow,
+		# but they're not.
+		for panel in self.__detachedPanels :
+			panel.setDisplayTransform( self.displayTransform() )
+
 	def __visibilityChanged(self, widget) :
 
 		v = self.visible()
@@ -207,11 +213,12 @@ class CompoundEditor( GafferUI.Editor ) :
 
 	def __parentChanged( self, widget ) :
 
-		# Make sure we have the correct keyboard shortcut listeners
+		# If we created any detached panels before we were
+		# in a ScriptWindow, connect them to the ScriptWindow now.
 		scriptWindow = self.ancestor( GafferUI.ScriptWindow )
 		if scriptWindow is not None :
 			for panel in self._detachedPanels() :
-				scriptWindow.menuBar().addShortcutTarget( panel )
+				panel.connectToScriptWindow( scriptWindow )
 
 	def __repr__( self ) :
 
@@ -919,6 +926,13 @@ class _DetachedPanel( GafferUI.Window ) :
 		if self.__windowState :
 			_restoreWindowState( self, self.__windowState )
 
+	def connectToScriptWindow( self, scriptWindow ) :
+
+		scriptWindow.menuBar().addShortcutTarget( self )
+		self.setTitle( scriptWindow.getTitle() )
+		scriptWindow.titleChangedSignal().connect( Gaffer.WeakMethod( self.__scriptWindowTitleChanged ) )
+		self.setDisplayTransform( scriptWindow.displayTransform() )
+
 	# A detached panel is considered empty if it has a single split with no editors.
 	def isEmpty( self ) :
 
@@ -941,6 +955,10 @@ class _DetachedPanel( GafferUI.Window ) :
 	# Required for editor path introspection
 	def _splitContainer( self ) :
 		return self.__splitContainer
+
+	def __scriptWindowTitleChanged( self, window, title ) :
+
+		self.setTitle( title )
 
 ## An internal eventFilter class managing all tab drag-drop events and logic.
 # Tab dragging is an exception and is implemented entirely using mouse-move